Move most classes from model package to dto
