+++ src/main/java/org/developerden/codosseum/dto/PlayerGameResult.java
@@ -15,14 +15,12 @@
  *
  */
 
-package org.developerden.codosseum.model;
+package org.developerden.codosseum.dto;
 
 import io.soabase.recordbuilder.core.RecordBuilder;
-import jakarta.annotation.Nonnull;
 
 @RecordBuilder
-public record Player(
-    @Nonnull
-    String name
+public record PlayerGameResult(
+    double score
 ) {
 }
